Fixes issue #8, incorrect headings (and actual missing data for main fire and some object fires) in "normal" spreadsheet output file.
Fixes part of issue #5, correcting data changed flag for filter efficiency. Rest of #5 (GUI dying on some values of filter efficiency input) awaits verification of problem.

Changed location of some offspring windows in GUI to be at the center of parent window rather than center of screen.  This was an issue only for computers with multiple displays.