Updated gauge heat flux output to be incident heat flux (assumes emissivity of 1 for the "gauge" instead of the emissivity given for the material).  Target actually still only absorbs radiation based on emissivity so no change in target temperature.
Updated calculation of layer absorbtivities for targets to be consistent with current calculations for surfaces.  Targets were still using the old assumed absorbtivities of 0.5 for the upper layer and 0.1 for the lower layer.  This had been updated years ago for surfaces to be based on a calculated value that depends on the CO2 and H2O concentrations in the layers.  Targets seem to have been overlooked at that time.  This change results in a bit quicker temperature rise for the targets early in the fire when the layers are relatively clear.  Result is a bit higher target temepratures for some scenarios.  Data for NRC verification tests matches FDS results better with this update.