Added chemie back into fire.f now that it's working and has been tested, verified, and validated.
Updated printout for fire specification to be consistent with new chemistry inputs.

Corrected undefined variable for pyrolysis temperature for object fires.  Like the main fire, this is just defined to be the default reference temperature.

Updated validation scatter plots and verified that results have not changed as a result of this update.